fix(rmm): fix MISRA C 2012 violations
Fixes MISRA C 2012 Rules 2.1, 2.2, 8.3, 8.4,
9.1, 10.3, 10.6, 14.2, 16.5, 21.2, 21.15 and
21.18 violations.

+++ lib/attestation/src/attestation_token.c
@@ -331,5 +331,7 @@ int attest_realm_token_create(enum hash_algo algorithm,
 	QCBOREncode_CloseBstrWrap2(&(ctx->ctx.cbor_enc_ctx), false,
 				   &(ctx->ctx.signed_payload));
 
-	return ATTEST_TOKEN_ERR_SUCCESS;
+	/* coverity[misra_c_2012_directive_4_4_violation:SUPPRESS] */
+	/* NOLINTNEXTLINE(google-readability-casting) */
+	return (int)ATTEST_TOKEN_ERR_SUCCESS;
 }
